+++ src/loader/cmi.mli
@@ -96,3 +96,9 @@ val read_extension_constructor : env ->
 val read_exception : env ->
   Paths.Identifier.Signature.t -> Ident.t ->
   Types.extension_constructor -> Odoc_model.Lang.Exception.t
+
+#if defined OXCAML
+val read_jkind_annotation :
+  Parsetree.jkind_annotation option ->
+  Odoc_model.Lang.KindAnnotation.t
+#endif